Refactor Encounter Actions into a Reusable Component
- Create new `EncounterActions` component to centralize encounter-related actions
- Support both dropdown and standalone layouts
- Integrate with existing components like PatientInfoCard and QuickAccess
- Simplify action handling and improve code reusability
